+++ Assets/Prey.cs
@@ -9,7 +9,7 @@ internal class Prey : Entity
 {
 
 	//set prey entity to predetermined values
-	public Prey(float energyLevel, float foodLevel, float waterLevel, int maxOffsprings, float reproductionProb, int numOffsprings, float minReproductionEnergy, string name, int x, int y)
+	public Prey(float energyLevel, float foodLevel, float waterLevel, int maxOffsprings, int reproductionProb, int numOffsprings, float minReproductionEnergy, string name, int x, int y)
 	{
 		this.energyLevel = energyLevel;
 		this.foodLevel = foodLevel;
